+++ packages/zowe-explorer/__mocks__/vscode.ts
@@ -9,6 +9,57 @@
  *
  */
 
+export enum ViewColumn {
+    /**
+     * A *symbolic* editor column representing the currently active column. This value
+     * can be used when opening editors, but the *resolved* {@link TextEditor.viewColumn viewColumn}-value
+     * of editors will always be `One`, `Two`, `Three`,... or `undefined` but never `Active`.
+     */
+    Active = -1,
+    /**
+     * A *symbolic* editor column representing the column to the side of the active one. This value
+     * can be used when opening editors, but the *resolved* {@link TextEditor.viewColumn viewColumn}-value
+     * of editors will always be `One`, `Two`, `Three`,... or `undefined` but never `Beside`.
+     */
+    Beside = -2,
+    /**
+     * The first editor column.
+     */
+    One = 1,
+    /**
+     * The second editor column.
+     */
+    Two = 2,
+    /**
+     * The third editor column.
+     */
+    Three = 3,
+    /**
+     * The fourth editor column.
+     */
+    Four = 4,
+    /**
+     * The fifth editor column.
+     */
+    Five = 5,
+    /**
+     * The sixth editor column.
+     */
+    Six = 6,
+    /**
+     * The seventh editor column.
+     */
+    Seven = 7,
+    /**
+     * The eighth editor column.
+     */
+    Eight = 8,
+    /**
+     * The ninth editor column.
+     */
+    Nine = 9,
+}
+
 /**
  * A provider result represents the values a provider, like the [`HoverProvider`](#HoverProvider),
  * may return. For once this is the actual result type `T`, like `Hover`, or a thenable that resolves
@@ -237,6 +288,22 @@ export namespace window {
 
     export function registerFileDecorationProvider(provider: FileDecorationProvider) {
         return this;
+    }
+
+    export function createWebviewPanel(
+        viewType: string,
+        title: string,
+        showOptions: ViewColumn | { viewColumn: ViewColumn; preserveFocus?: boolean },
+        options?: any
+    ): any {
+        return {
+            onDidDispose: jest.fn(),
+            webview: {
+                asWebviewUri: jest.fn(),
+                postMessage: jest.fn(),
+                onDidReceiveMessage: jest.fn(),
+            },
+        };
     }
 
     /**