+++ packages/zowe-explorer/__tests__/__unit__/uss/AttributeView.unit.test.ts
@@ -1,0 +1,100 @@
+import { ExtensionContext } from "vscode";
+import { AttributeView } from "../../../src/uss/AttributeView";
+import { IZoweTree, IZoweUSSTreeNode, ZoweExplorerApi } from "@zowe/zowe-explorer-api";
+import { ZoweExplorerApiRegister } from "../../../src/ZoweExplorerApiRegister";
+
+describe("AttributeView unit tests", () => {
+    let view: AttributeView;
+    const context = { extensionPath: "some/fake/ext/path" } as unknown as ExtensionContext;
+    const treeProvider = { refreshElement: jest.fn(), refresh: jest.fn() } as unknown as IZoweTree<IZoweUSSTreeNode>;
+    const node = {
+        attributes: {
+            perms: "----------",
+        },
+        label: "example node",
+        fullPath: "/z/some/path",
+        getParent: jest.fn(),
+        getProfile: jest.fn(),
+        onUpdate: jest.fn(),
+    } as unknown as IZoweUSSTreeNode;
+    const updateAttributesMock = jest.fn();
+
+    beforeAll(() => {
+        jest.spyOn(ZoweExplorerApiRegister, "getUssApi").mockReturnValue({
+            updateAttributes: updateAttributesMock,
+        } as unknown as ZoweExplorerApi.IUss);
+        view = new AttributeView(context, treeProvider, node);
+    });
+
+    afterEach(() => {
+        node.onUpdate = jest.fn();
+    });
+
+    it("refreshes properly when webview sends 'refresh' command", () => {
+        // case 1: node is a root node
+        (view as any).onDidReceiveMessage({ command: "refresh" });
+        expect(treeProvider.refresh).toHaveBeenCalled();
+
+        // case 2: node is a child node
+        node.getParent = jest.fn().mockReturnValueOnce({ label: "parent node" } as IZoweUSSTreeNode);
+        (view as any).onDidReceiveMessage({ command: "refresh" });
+        expect(treeProvider.refreshElement).toHaveBeenCalled();
+
+        expect(node.onUpdate).toHaveBeenCalledTimes(2);
+    });
+
+    it("dispatches node data to webview when 'ready' command is received", () => {
+        (view as any).onDidReceiveMessage({ command: "ready" });
+        expect(view.panel.webview.postMessage).toHaveBeenCalledWith({
+            attributes: node.attributes,
+            name: node.fullPath,
+            readonly: false,
+        });
+    });
+
+    it("updates attributes when 'update-attributes' command is received", async () => {
+        // case 1: no attributes provided from webview (sanity check)
+        (view as any).onDidReceiveMessage({ command: "update-attributes" });
+        expect(updateAttributesMock).not.toHaveBeenCalled();
+
+        const attributes = {
+            owner: "owner",
+            group: "group",
+            perms: "-rwxrwxrwx",
+        };
+
+        // case 2: attributes provided from webview, pass owner/group as name
+        await (view as any).onDidReceiveMessage({
+            command: "update-attributes",
+            attrs: attributes,
+        });
+        expect(updateAttributesMock).toHaveBeenCalled();
+        expect(view.panel.webview.postMessage).toHaveBeenCalledWith({
+            updated: true,
+        });
+
+        // case 2: attributes provided from webview, pass owner/group as IDs
+        await (view as any).onDidReceiveMessage({
+            command: "update-attributes",
+            attrs: {
+                ...attributes,
+                owner: "1",
+                group: "9001",
+            },
+        });
+        expect(updateAttributesMock).toHaveBeenCalled();
+        expect(view.panel.webview.postMessage).toHaveBeenCalled();
+    });
+
+    it("handles any errors while updating attributes", async () => {
+        updateAttributesMock.mockRejectedValueOnce(new Error("Failed to update attributes"));
+        await (view as any).onDidReceiveMessage({
+            command: "update-attributes",
+            attrs: {},
+        });
+        expect(updateAttributesMock).toHaveBeenCalled();
+        expect(view.panel.webview.postMessage).toHaveBeenCalledWith({
+            updated: false,
+        });
+    });
+});